cast int32 to float32

+++ mobilenet/mobilenet.py
@@ -160,7 +160,7 @@ def sparse_top_k_categorical_accuracy(y_true, y_pred, k=5):
        len(tf.keras.backend.int_shape(y_pred)))):
     y_true = tf.squeeze(y_true, [-1])
 
-  y_true = tf.cast(y_true, 'int32')
+  y_true = tf.cast(y_true, 'float32')
   return tf.nn.in_top_k(y_pred, y_true, k)
